+++ kitty/glfw-wrapper.h
@@ -2544,6 +2544,10 @@ typedef pid_t (*glfwWaylandCompositorPID_func)(void);
 GFW_EXTERN glfwWaylandCompositorPID_func glfwWaylandCompositorPID_impl;
 #define glfwWaylandCompositorPID glfwWaylandCompositorPID_impl
 
+typedef double (*glfwGetWaylandPrimaryMonitorFractionalScale_func)(void);
+GFW_EXTERN glfwGetWaylandPrimaryMonitorFractionalScale_func glfwGetWaylandPrimaryMonitorFractionalScale_impl;
+#define glfwGetWaylandPrimaryMonitorFractionalScale glfwGetWaylandPrimaryMonitorFractionalScale_impl
+
 typedef void (*glfwConfigureMomentumScroller_func)(double, double, double, unsigned);
 GFW_EXTERN glfwConfigureMomentumScroller_func glfwConfigureMomentumScroller_impl;
 #define glfwConfigureMomentumScroller glfwConfigureMomentumScroller_impl